feat(InputGroup): migrate to Tailwind
